\name{genUniform}
\alias{genUniform}
\title{Get uniformly a spin state}
\usage{
  genUniform(n)
}
\arguments{
 \item{n}{dummy argument}
}
\value{
  Returns randomly 1 or -1 from uniform distribution.
}
\description{
  Generate a single spin state from uniform distribution.
}
\examples{
 genUniform()
}
\author{
  Mehmet Suzen <mehmet.suzen@physics.org>
}
